YP-3: Исправил локальные замечания командного ревью
